clean up check snapshots files
